% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeb.R
\name{jeb}
\alias{jeb}
\title{Jeb}
\usage{
jeb(image = "default", lookup = "GADM3", country = "USA",
  level = 1, bg_col = "goldenrod", name = NULL)
}
\arguments{
\item{image}{The image to overlay. Default to Jeb! Bush celebrating, otherwise, a file path to (preferably) a .png file}

\item{lookup}{The method used to find the shapefile to use as the electoral background. Defaults to GADM}

\item{country}{The GADM code for the country shapefiles to download from GADM. If not using GADM, searches the environment for a matching object}

\item{level}{The GADM level for the country shapefiles to download from GADM. Requires GADM to be the lookup method}

\item{bg_col}{The background fill colour for the electoral shapefiles. Defaults to a Goldenrod shape.}

\item{name}{The name for the country on the final plot title. If using GADM, will find this from the NAME_0 column of the downloaded shape}
}
\description{
Produce an election victory map overlaid with a picture (e.g. of Jeb! Bush)
}
\details{
Generates a plot of an image over an electoral map. E.g. quickly and easily allows for Jeb! meme style maps to be generated
}
\examples{
plot <- Jeb::jeb(image = "default",
 lookup = "GADM3", country = "Sweden", level = 2,
  bg_col = "Goldenrod", name = "Swedish Federal")

}
