% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeb.R
\name{jeb}
\alias{jeb}
\title{Jeb}
\usage{
jeb(image = "default", lookup = "GADM", shape_code = "USA",
  level = 1, bg_col = "goldenrod")
}
\arguments{
\item{image}{The image to overlay. Default to Jeb! Bush celebrating, otherwise, a file path to (preferably) a .png file}

\item{lookup}{The method used to find the shapefile to use as the electoral background. Defaults to GADM}

\item{shape_code}{The GADM code for the country shapefiles to download from GADM. Requires GADM to be the lookup method}

\item{bg_col}{The background fill colour for the electoral shapefiles. Defaults to a Goldenrod shape.}

\item{geocode}{The GADM level for the country shapefiles to download from GADM. Requires GADM to be the lookup method}
}
\description{
Produce an election victory map overlaid with a picture (e.g. of Jeb! Bush)
}
\details{
Generates a plot of an image over an electoral map. E.g. quickly and easily allows for Jeb! meme style maps to be generated
}
\examples{

}
